Тест с полной матрицей и с RT и IT, равными 32 бита каждый.

+++ src/prec.jl
@@ -6,7 +6,7 @@ module prec
 export IT, RT
 
 # Определение типа целых чисел
-IT = Int64
+IT = Int32
 
 # Определение типа вещественных чисел
 RT = Float64
